refactor(/controller): test 코드의 패키지 경로를 main과 일치하도록 변
