+++ tomcat/src/main/java/com/techcourse/controller/LoginController.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 import org.apache.catalina.session.Session;
 import org.apache.catalina.session.SessionManager;
 import org.apache.coyote.exception.UnauthorizedException;
-import org.apache.coyote.handler.AbstractController;
+import org.apache.coyote.controller.AbstractController;
 import org.apache.http.HttpCookie;
 import org.apache.http.request.HttpRequest;
 import org.apache.http.response.HttpResponse;
