+++ tomcat/src/main/java/org/apache/coyote/controller/StaticResourceController.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.handler;
+package org.apache.coyote.controller;
 
 import java.nio.file.Path;
