+++ tomcat/src/test/java/org/apache/coyote/HandlerMappingTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.apache.coyote.exception.NotFoundException;
 import org.apache.coyote.exception.UnauthorizedException;
-import org.apache.coyote.handler.Controller;
-import org.apache.coyote.handler.StaticResourceController;
+import org.apache.coyote.controller.Controller;
+import org.apache.coyote.controller.StaticResourceController;
 import org.apache.http.HttpMethod;
 import org.apache.http.HttpVersion;
 import org.apache.http.header.HttpHeaders;
