+++ tomcat/src/test/java/org/apache/coyote/controller/StaticResourceControllerTest.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package org.apache.coyote.controller;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -7,7 +7,6 @@ import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
-import org.apache.coyote.handler.StaticResourceController;
 import org.apache.http.header.HttpHeaderName;
 import org.apache.http.request.HttpRequest;
 import org.apache.http.request.RequestLine;
